requirements.txt.txt
requests
beautifulsoup4
lxml
pandas
SQLAlchemy
PyMySQL
python-dotenv
openpyxl
